requirements.txt
Django==5.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

channels==4.1.0
channels-redis==4.2.0

python-dotenv==1.0.1
django-redis==5.4.0
redis==5.0.8

daphne==4.1.2
Pillow==10.4.0
